Refactor GetDatabasesToCollect to only take config instead of a server
This unnecessary passed a pointer to the server struct around, when we
can be more specific to just pass a config around instead.

In passing move the isCloudInternalDatabase next to the function since
it is the only caller.
