+++ runner/generate_helper_sql.go
@@ -56,7 +56,7 @@ func GenerateStatsHelperSql(ctx context.Context, server *state.Server, globalCol
 	}
 
 	output := strings.Builder{}
-	for _, dbName := range postgres.GetDatabasesToCollect(server, databases) {
+	for _, dbName := range postgres.GetDatabasesToCollect(server.Config, databases) {
 		output.WriteString(fmt.Sprintf("\\c %s\n", pq.QuoteIdentifier(dbName)))
 		output.WriteString("CREATE SCHEMA IF NOT EXISTS pganalyze;\n")
 		output.WriteString(fmt.Sprintf("GRANT USAGE ON SCHEMA pganalyze TO %s;\n", server.Config.GetDbUsername()))
@@ -87,7 +87,7 @@ func GenerateExplainAnalyzeHelperSql(ctx context.Context, server *state.Server, 
 	}
 
 	output := strings.Builder{}
-	for _, dbName := range postgres.GetDatabasesToCollect(server, databases) {
+	for _, dbName := range postgres.GetDatabasesToCollect(server.Config, databases) {
 		output.WriteString(fmt.Sprintf("\\c %s\n", pq.QuoteIdentifier(dbName)))
 		output.WriteString("CREATE SCHEMA IF NOT EXISTS pganalyze;\n")
 		output.WriteString(fmt.Sprintf("GRANT USAGE ON SCHEMA pganalyze TO %s;\n", pq.QuoteIdentifier(server.Config.GetDbUsername())))
